requirements.txt
fastapi==0.115.8
httpx==0.28.1
prometheus-fastapi-instrumentator==7.0.2
pyyaml==6.0.2
uvicorn==0.34.0